requirements.txt
fastapi==0.100.0
greenlet==2.0.2
pydantic==2.7.0
pydantic-settings==2.0.1
requests==2.31.0
uvicorn==0.22.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
urllib3==2.0.4
gunicorn==21.2.0
fastapi-cache==0.1.0
pytz==2024.1
aiocache
itsdangerous
openai
pycountry